Dziala ( pamietamy ze int nie dzieli sie rzeczywiscie)

+++ Rouge/KlasyPodstawowe/Player.cs
@@ -366,8 +366,8 @@ namespace Rouge
                 
                 int totalDamage = attackRight.Damage + attackLeft.Damage;
                 room._enemiesMap[(player.SelectedEnemy.Y, player.SelectedEnemy.X)].CurrentHealth -= totalDamage;
-                if (room._enemiesMap[(player.SelectedEnemy.Y, player.SelectedEnemy.X)].CurrentHealth /
-                    room._enemiesMap[(player.SelectedEnemy.Y, player.SelectedEnemy.X)].EnemyStats.Health <= 0.2)
+                if ((double)room._enemiesMap[(player.SelectedEnemy.Y, player.SelectedEnemy.X)].CurrentHealth /
+                    (double)room._enemiesMap[(player.SelectedEnemy.Y, player.SelectedEnemy.X)].EnemyStats.Health <= 0.2)
                 {
                     room._enemiesMap[(player.SelectedEnemy.Y, player.SelectedEnemy.X)].Behavior = new FearfulBehavior();
                     GameDisplay.Instance?.AddLogMessage(
